+++ aftiktuna-rust/CHANGELOG.md
@@ -17,7 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Color variants for scarvie feathers
 - New command "ask about \<item\>" that can be used in stores
 - Dialogue system addition: npc background id
-- New interaction with recruitable npc
+- New interactions with recruitable npcs
 - New data files `fauna.json` and `species.json`
 - New data file `character_names.json` for random selection of non-aftik character names